Fix G++ warnings

+++ src/util.h
@@ -27,6 +27,7 @@
 
 __M_BEGIN_DECLS
 
+#include <stdint.h>
 #include <stdio.h>
 #if !defined(_MSC_VER) && !defined(__amigaos__)
 #include <unistd.h>
@@ -100,9 +101,9 @@ void fputd(int src, FILE *fp);
 CORE_LIBSPEC long ftell_and_rewind(FILE *f);
 
 CORE_LIBSPEC void rng_seed_init(void);
-unsigned long long rng_get_seed(void);
-void rng_set_seed(unsigned long long seed);
-unsigned int Random(unsigned long long range);
+uint64_t rng_get_seed(void);
+void rng_set_seed(uint64_t seed);
+unsigned int Random(uint64_t range);
 
 CORE_LIBSPEC void add_ext(char *src, const char *ext);
 CORE_LIBSPEC int get_ext_pos(const char *filename);
